Move nodetool into metric package for easy access to the test helpers
