+++ cassandra-operator/pkg/metrics/ready.go
@@ -1,11 +1,10 @@
-package nodetool
+package metrics
 
 import (
 	"log"
 	"time"
 
 	"github.com/sky-uk/cassandra-operator/cassandra-operator/pkg/cluster"
-	"github.com/sky-uk/cassandra-operator/cassandra-operator/pkg/metrics"
 )
 
 type localGatherer struct {
@@ -19,13 +18,13 @@ type Nodetool struct {
 	cluster *cluster.Cluster
 }
 
-func New(cluster *cluster.Cluster) *Nodetool {
+func NewNodetool(cluster *cluster.Cluster) *Nodetool {
 	return &Nodetool{cluster: cluster}
 }
 
 func (n *Nodetool) IsLocalNodeReady() (bool, error) {
 	j := &localGatherer{}
-	gatherer := metrics.NewGatherer(j, &metrics.Config{
+	gatherer := NewGatherer(j, &Config{
 		RequestTimeout: 20 * time.Second,
 	})
 	status, err := gatherer.GatherMetricsFor(n.cluster)
